[CMSA-284] [웅진] demo-api 무정지 배포 cicd 가이드 문서 작성
[Resolution]
demo-api 가이드 문서 초안 작성

[DoD] demo-api Blue-Green CICD 가이드 문서 작성 및 가이드 및 고객 확인

[Team] Productivity Engineering
[Company] BespinGlobal
