+++ .gitignore
@@ -27,3 +27,4 @@ hs_err_pid*
 
 # ehcache
 /tmp/ehcache
+/demo-api.iml
