+++ pom.xml
@@ -1,18 +1,18 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
         <version>2.2.6.RELEASE</version>
-        <relativePath />
+        <relativePath/>
     </parent>
     <groupId>com.bespin.demo</groupId>
-    <artifactId>backend-demo</artifactId>
+    <artifactId>demo-api</artifactId>
     <version>1.0.0-SNAPSHOT</version>
-    <name>backend-demo</name>
-    <description>Demo project for Spring Boot</description>
+    <name>Demo API Application</name>
+    <description>Demo application implemented by Spring Boot</description>
 
     <properties>
         <java.version>1.8</java.version>
