+++ src/main/java/com/bespin/demo/message/DemoRequest.java
@@ -1,0 +1,4 @@
+package com.bespin.demo.message;
+
+public class DemoRequest {
+}
