+++ src/main/java/com/bespin/demo/repository/DemoMapper.java
@@ -1,0 +1,4 @@
+package com.bespin.demo.repository;
+
+public class DemoMapper {
+}
