+++ src/main/java/com/bespin/demo/support/Utils.java
@@ -1,0 +1,4 @@
+package com.bespin.demo.support;
+
+public class Utils {
+}
