+++ src/main/resources/scripts/afterInstall.sh
