+++ src/main/resources/scripts/applicationStart.sh
