+++ src/main/resources/scripts/applicationStop.sh
