+++ src/main/resources/scripts/beforeInstall.sh
