+++ src/test/java/com/bespin/demo/AbstractTests.java
@@ -1,0 +1,17 @@
+package com.bespin.demo;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+public abstract class AbstractTests {
+
+    private final ObjectMapper JACKSON_MAPPER = new ObjectMapper();
+
+    protected String toJson(final Object obj) {
+        try {
+            return JACKSON_MAPPER.writeValueAsString(obj);
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+}
